+++ packages/orq-rc/RELEASES.md
@@ -3579,3 +3579,13 @@ Based on:
 - [typescript v3.13.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.3 - packages/orq-rc
+
+## 2025-09-10 08:54:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.615.0 (2.698.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.4 - packages/orq-rc